+++ Game/Players/Concrete/CpuPlayer.cs
@@ -4,7 +4,7 @@ namespace Game.Players.Concrete
 {
 	internal class CpuPlayer : BasePlayer
 	{
-		public CpuPlayer(Guid id, string name, float balance, int number) : base(id, name, balance, number)
+		public CpuPlayer(Guid id, string name, double balance, int number) : base(id, name, balance, number)
 		{
 		}
 	}
