+++ Game/Services/Abstract/IMessageDisplayService.cs
@@ -5,10 +5,10 @@ namespace Game.Services.Abstract
 {
 	public interface IMessageDisplayService
 	{
-		public void DisplayWelcomeMessage(BasePlayer? player, float ticketPrice);
+		public void DisplayWelcomeMessage(BasePlayer? player, double ticketPrice);
 		public void DisplayPurchaseTicketsMessage(int allowedTicketsAmountToPurchaseMinimum, int allowedTicketsAmountToPurchaseMaximum);
 		public void DisplayCpuPlayerPurchaseTicketsMessage(List<BasePlayer>? players);
 		public void DisplayWinners(List<WinnerDto> winners);
-		public void DisplayHouseProfit(float profit);
+		public void DisplayHouseProfit(double profit);
 	}
 }
