+++ Game/Tickets/Concrete/Ticket.cs
@@ -4,7 +4,7 @@ namespace Game.Tickets.Concrete
 {
 	internal class Ticket : BaseTicket
 	{
-		public Ticket(Guid id, float price) : base(id, price)
+		public Ticket(Guid id, double price) : base(id, price)
 		{
 		}
 	}
